Fixed a race condition with dictionaries.
For clones with same score, the order of these clones varied between executions. This produced that the results of the program where kind of random (for example, nars-console update filter did not produce the same results for the same filter). I solved this issue by always ordering alphabetically the clones when they have the same score and in the original PClone list with no scores (for arwork substitution to work deterministically).

Some other minor changes here and there.